Capstone Project/requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.2
cycler==0.12.1
fonttools==4.59.1
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.5
numpy==2.2.6
packaging==25.0
pandas==2.3.2
pillow==11.3.0
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
seaborn==0.13.2
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
